+++ src/server/routers/player/liveGame/router.ts
@@ -1,9 +1,10 @@
 import { router } from "@/server/trpc"
 
-import { getAllData, getDuration, getMode } from "./endpoints"
+import { getAllData, getDuration, getMode, getPlayers } from "./endpoints"
 
 export const liveGameRouter = router({
 	getMode,
 	getDuration,
+	getPlayers,
 	getAllData,
 })